refactor(hush-cli): extract pack from pkg_cli
Move cmd_pkg_pack plus its archive/validation helpers
(archive_file_name, copy_dir_recursive_excluding_cpkg,
pack_source_dir_without_embedded_archives, validate_pack_contents)
into pkg_cli/pack.rs. cmd_pkg_pack and the two helpers reused by the
publish handler widened to pub(super); validate_pack_contents is
pub(super) for tests; copy_dir_recursive_excluding_cpkg stays private.
PkgType import in mod.rs is now test-only (#[cfg(test)]).
